Delete top-level share-parcel schema
This schema is completely broken. It's `allOf` a `base-statement` and `share-parcel-data`, and `base-statement` has a `data` property which is a _`financial-payment-data-object`_ - clearly nonsensical. So the only way a json document would validate successfully against this schema is if it had a `data` property whose elements matched _both_ `share-parcel-data` and `financial-payment-data-object`, which doesn't seem desirable at all, and probably impossible.

Given it's unlikely any data has ever been validated against this schema (otherwise) - TBC

* https://github.com/openc/openc-schema/commit/ba85dfb4a83e3d198f8e2367b5d90f8585eb9bb4
* https://github.com/openc/openc-schema/commit/976c64e94ff664faea1801a884175daa1ab650f2
* https://github.com/openc/openc-schema/commit/9d831cc2b5d2bfe6864373434967f3db09ea4f9f
* https://github.com/openc/openc-schema/commit/997b73dd16076a90f3267dc50f75eaff859229d5
* https://github.com/openc/openc-schema/commit/e084b9d570350613cae065b16de8b3ee28dc4ffa
* https://github.com/openc/openc-schema/commit/5e3ba4b5e97427822a6d214a9130d1c2b89b03d9