Delete test fixture that is wrong and in wrong place
It's in `spec/sample-data/` which means it doesn't get picked up by the specs - it needs to be in `spec/sample-data/invalid/filing/`. But also it's wrong anyway because these days an entity is required to have _either_ a `name` or `company_number`. This one has a `company_number` so is valid, not invalid.